Match 55 functions at 97-99.99% (100% each, zero regressions)
+12,692 bytes matched code (+0.33%)

Matched functions:
- ftKb_SpecialN/Init/SpecialNNs/SpecialNPk/SpecialNYs/SpecialNZd/SpecialNFx: 35 Kirby copy ability functions
- grIceMt, grFourside, grKongo, grVenom: 5 stage functions
- ftCo_800D105C, ftCo_CaptureWaitHi_Anim: 2 common fighter functions
- it_802E5F8C: 1 item function
- fn_80198C60, fn_8030110C, fn_803011EC: 3 misc functions
- HSD_SisLib_803A6048, hsd_803AC340: 2 sysdolphin functions
- un_80312834, un_80313358, un_803062EC: 3 trophy/toy functions
- ftKb_EatTurn_Anim: 1 Kirby eat function

+++ src/melee/gm/gm_1BA8.c
@@ -1472,7 +1472,7 @@ void gm_801BEF84(MinorScene* arg)
 
 #pragma push
 #pragma dont_inline on
-void gm_801BEFA4(s8 ckind)
+void gm_801BEFA4(int ckind)
 {
     gm_8049C178[0] = ckind;
 }
